refactor(core): move resolver helpers into resolution folder
