+++ crates/tsz-core/src/lib.rs
@@ -207,13 +207,13 @@ pub mod config;
 // Module Resolution Infrastructure (non-wasm targets only - requires file system access)
 #[cfg(not(target_arch = "wasm32"))]
 pub mod module_resolver;
-#[cfg(not(target_arch = "wasm32"))]
-mod module_resolver_helpers;
+mod resolution;
 #[cfg(not(target_arch = "wasm32"))]
 pub use module_resolver::{
     ModuleExtension, ModuleLookupError, ModuleLookupOutcome, ModuleLookupRequest,
     ModuleLookupResult, ModuleResolver, ResolutionFailure, ResolvedModule,
 };
+pub(crate) use resolution::helpers as module_resolver_helpers;
 
 // Import/Export Tracking
 pub mod imports;
